backend/requirements.txt
Flask
pillow
Flask-Cors

opencv-python
numpy